tests/wpt/web-platform-tests/wake-lock/wakelockrequest-is-independent.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>WakeLockRequest object is independent</title>
<link rel="help" href="https://w3c.github.io/wake-lock/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async t => {
  const wakeLock1 = await navigator.getWakeLock("screen");
  const wakeLock2 = await navigator.getWakeLock("screen");
  const request1 = wakeLock1.createRequest();
  const request2 = wakeLock2.createRequest();
  assert_not_equals(request1, request2);
}, "Test that the WakeLockRequest object is independent.");
</script>
